requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
oci==2.110.1
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
cryptography==41.0.4
pytest==7.4.3
black==23.10.1
isort==5.12.0
mypy==1.6.1
python-jose==3.3.0
passlib==1.7.4
